Finds first location visited twice
This took me a lot of steps, and it probably isn't the cleanest code,
but it is working okay. I think I would have done better with many tests
along the way. I'm happy with some parts, but not all of them. I'd
really like to clean up some wording, but I think this is good to move
on to the next day.

Amos King @adkron <amos@binarynoggin.com>